+++ src/pool/dynamic.rs
@@ -134,6 +134,7 @@ pub fn create_dynamic_pool(
             .unwrap_or(config.general.idle_timeout.as_millis()),
         config.general.server_idle_check_timeout.as_millis(),
         config.general.connect_timeout.as_std(),
+        config.general.query_wait_timeout.as_std(),
         pool_mode == PoolMode::Session,
         fallback_state,
     );